修改bug

+++ iceking-core/src/main/java/iceking/base/mapper/BaseMapper.java
@@ -1,13 +1,13 @@
 package iceking.base.mapper;
 
 public interface BaseMapper<T> {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(T t);
 
     int insertSelective(T t);
 
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(T t);
